Renamed mock dir to mockpb
