+++ src/tests/mockpb/pybitmessage/singleton.py
